Code improvement
1. To make it easier to distinguish pointers from variables containing
addresses during address calculations, now uintptr_t type is used
instead of uint32_t in address calculations, also now variables containing
address have "_addr" at the end of their name, while pointers have "_ptr"
at the end of their name.

+++ kernel/arch/i386/multiboot/multiboot_utils.h
@@ -8,6 +8,6 @@
 #include <stdio.h>
 #include "multiboot2.h"
 
-void multiboot_print_mbi_info(uint32_t mbi_addr);
+void multiboot_print_mbi_info(uintptr_t mbi_addr);
 
 #endif
